feat(user): 实现用户注册功能并优化用户实体类
- 添加 UserController 控制器，提供用户注册接口
- 创建 UserRegisterRequest DTO，封装用户注册请求参数
- 修改 User 实体类中的 isDelete 字段类型为 Byte
- 在 UserService 中新增用户注册及密码加密方法
- 实现用户注册业务逻辑，包括参数校验、账号重复检查和密码加密
- 使用 MD5 加密算法对用户密码进行加密处理

+++ backEnd/src/main/java/com/cloud/picture/space/backend/model/entity/User.java
@@ -95,7 +95,7 @@ public class User implements Serializable {
      */
     //逻辑删除标记: @TableLogic 是 MyBatis-Plus 框架提供的注解，用于实现逻辑删除功能
     @TableLogic
-    private Integer isDelete;
+    private Byte isDelete;
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
